nothing useful

+++ Labulya/Labulya/lec.cpp
@@ -3,10 +3,10 @@
 
 lexeme::lexeme(char * name, char * type, char * data, int line, int startposition, int priority)
 {
-	_name = (char*)heap.get_mem(sizeof(char) * strlen(name) + 1);
+	_name = (char*)heap.get_mem(sizeof(char) * (int)strlen(name) + 1);
 	strcpy_s(_name, strlen(name) + 1, name);
 	_type = GetType(type);
-	_data = (char*)heap.get_mem(sizeof(char) * strlen(data) + 1);
+	_data = (char*)heap.get_mem(sizeof(char) * (int)strlen(data) + 1);
 	strcpy_s(_data, strlen(data) + 1, data);
 	_line = line;
 	_startposition = startposition;
@@ -15,10 +15,10 @@ lexeme::lexeme(char * name, char * type, char * data, int line, int startpositio
 
 lexeme::lexeme(char * name, int type, char * data, int line, int startposition, int priority)
 {
-	_name = (char*)heap.get_mem(sizeof(char) * strlen(name) + 1);
+	_name = (char*)heap.get_mem(sizeof(char) * (int)strlen(name) + 1);
 	strcpy_s(_name, strlen(name) + 1, name);
 	_type = type;
-	_data = (char*)heap.get_mem(sizeof(char) * strlen(data) + 1);
+	_data = (char*)heap.get_mem(sizeof(char) * (int)strlen(data) + 1);
 	strcpy_s(_data, strlen(data) + 1, data);
 	_line = line;
 	_startposition = startposition;
@@ -28,7 +28,7 @@ lexeme::lexeme(char * name, int type, char * data, int line, int startposition, 
 void lexeme::ToMass(lexeme* name, int type, lexeme* val, int rank)
 {
 	heap.free_mem(_name);
-	_name = (char*)heap.get_mem(sizeof(char) * (strlen(name->Data()) + 1));
+	_name = (char*)heap.get_mem(sizeof(char) * ((int)strlen(name->Data()) + 1));
 	strcpy_s(_name, (strlen(name->Data()) + 1), name->Data());
 	_type = type;
 	heap.free_mem(_data);
@@ -79,7 +79,7 @@ bool lexeme::SetNumber(double value)
 void lexeme::DataChange(char* newdata)
 {
 	heap.free_mem(_data);
-	_data = (char*)heap.get_mem(sizeof(char) * strlen(newdata) + 1);
+	_data = (char*)heap.get_mem(sizeof(char) * (int)strlen(newdata) + 1);
 	strcpy_s(_data, strlen(newdata) + 1, newdata);
 }
